+++ src/main/java/community/community/model/User.java
@@ -2,7 +2,7 @@ package community.community.model;
 
 public class User {
 
-  private long id;
+  private Integer id;
   private String accountId;
   private String name;
   private String token;
@@ -10,57 +10,41 @@ public class User {
   private long gmtModified;
 
 
-  public long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
-
-
   public String getAccountId() {
     return accountId;
   }
-
   public void setAccountId(String accountId) {
     this.accountId = accountId;
   }
-
-
   public String getName() {
     return name;
   }
-
   public void setName(String name) {
     this.name = name;
   }
-
-
   public String getToken() {
     return token;
   }
-
   public void setToken(String token) {
     this.token = token;
   }
-
-
   public long getGmtCreate() {
     return gmtCreate;
   }
-
   public void setGmtCreate(long gmtCreate) {
     this.gmtCreate = gmtCreate;
   }
-
-
   public long getGmtModified() {
     return gmtModified;
   }
-
   public void setGmtModified(long gmtModified) {
     this.gmtModified = gmtModified;
   }
-
 }
